Refactor code base as per latest Create React App
