--- src/index.css
@@ -1,37 +1,0 @@
-h1 { margin-bottom: 0 }
-
-.author { margin-left: 2em; }
-
-p {
-  max-width: 80em;
-}
-
-body {
-  font: 14px "Century Gothic", Futura, sans-serif;
-  margin: 20px;
-  margin-right:40px;
-}
-
-.form {
-  margin-bottom: 1em;
-  float: left;
-}
-
-.viewport {
-  margin-left: 16em;
-}
-
-.grid-container {
-  margin-bottom: 1em;
-}
-
-table {
-  border-collapse: collapse;
-}
-
-td {
-  background: #fff;
-  border: 1px solid #999;
-  height: 34px;
-  width: 34px;
-}
