+++ src/styles/App.css
@@ -1,0 +1,22 @@
+h1 {
+  margin-bottom: 0;
+}
+
+.author {
+  margin-left: 2em;
+}
+
+p {
+  max-width: 80em;
+}
+
+body {
+  font: 14px "Century Gothic", Futura, sans-serif;
+  margin: 20px;
+  margin-right: 40px;
+}
+
+.form {
+  margin-bottom: 1em;
+  float: left;
+}
