+++ src/styles/Viewport.css
@@ -1,0 +1,7 @@
+.viewport {
+  margin-left: 16em;
+}
+
+.grid-container {
+  margin-bottom: 1em;
+}
